Update SvgExeLauncher.java
Made minor improvements to allow for passing command line arguments to the main class within the self-extracting jar file, as well as to allow the creation of jar files with spaces in the file name (this means that you can no longer create jar files that contain the escape sequence "%20" in the file name). Also fixed a minor bug that prevented arguments in the "Program Arguments" field from being sent to the main class. Any arguments passed in via the command line will be passed to the main class after any arguments entered into the "Program Arguments" field.