Renamed SvgExe to JarMatey
This is a pretty big change, but it's mostly cosmetic. I've renamed
SvgExe to JarMatey.

I've also decided to un-mavenize the project, since we only had one
dependency and we only use that dependency in a single line of code, and
maintaining a local repository for that seems like overkill.